% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-pars.R
\name{sample_parameters}
\alias{sample_parameters}
\title{Sample LDA Topic Model Parameters}
\usage{
sample_parameters(alpha, beta, num_documents)
}
\arguments{
\item{alpha}{Numeric vector of prior parameter for topics over documents.}

\item{beta}{Numeric vector of prior parameter for words over topics}

\item{num_documents}{Number of documents to generate parameters for}
}
\value{
Returns a list with theta and phi matrices. Theta represents
  probabilities of topics over documents. Phi represents probabilities of 
  words over topics.
}
\description{
Sample parameters simulating a topic model from Dirichlet
  distributions.
}
\note{
Number of topics is inferred from the length of \code{alpha}.
}
