% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmsamples-package.R
\docType{package}
\name{tmsamples-package}
\alias{tmsamples-package}
\title{Simulate a Term Frequency Matrix From a Topic Model}
\description{
Generate synthetic corpora from the parameters of a probabilistic topic
  model. Used for Tommy Jones's dissertation research. Unlikely to ever be
  put in a package repository.
}
\keyword{internal}
