% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-pars.R
\name{generate_zipf}
\alias{generate_zipf}
\title{Generate a Vector Following Zipf's Law.}
\usage{
generate_zipf(vocab_size, magnitude, zipf_par = 1.07)
}
\arguments{
\item{vocab_size}{Length of the returned vector.}

\item{magnitude}{Magnitude, sum total of the returned vector.}
}
\value{
Returns a numeric vector of length \code{vocab_size} whose entries
  have the shape described by Zipf's law and sum to \code{magnitude}.
}
\description{
Generate a vector following Zipf's law. Can be used as a
  parameter for a Dirichlet distribution downstream.
}
\examples{
z <- generate_zipf(
       10000,
       100,
       1.07
     )
plot(log10(seq_along(z)), log10(z))
}
