% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_documents.R
\name{sample_documents}
\alias{sample_documents}
\title{Sample a DTM or TCM based on hierarchical multinomial parameters}
\usage{
sample_documents(theta, phi, doc_lengths, verbose = TRUE, threads = 1)
}
\arguments{
\item{theta}{Matrix where each row is a document, each column is a topic and
the i,j entries represent the probability of sampling topic j from doc i.}

\item{phi}{Matrix where each row is a topic and each column is a token. The
i,j entries represent the probability of sampling token j from topic i.}

\item{doc_lengths}{Integer vector of document lengths. Must be the same length
as rows of \code{theta}.}

\item{verbose}{Boolean; do you want to print a simple progress bar out to the console?}

\item{threads}{Integer number of parallel threads, defaults to 1.}
}
\value{
Returns a sparse matrix whose rows represent documents, columns
  represent tokens and i,j entries are the count of token j in document i.
}
\description{
Sample a sparse DTM or TCM based on hierarchical multinomial paramers.
}
\examples{
library(tmsamples)

Nk <- 4
Nd <- 50
Nv <- 1000

alpha <- rgamma(Nk, 0.5)

beta <- generate_zipf(vocab_size = Nv, magnitude = 500, zipf_par = 1.1)

pars <- sample_parameters(alpha, beta, Nd)

doc_lengths <- rpois(Nd, 50)

dtm <- sample_documents(
  theta = pars$theta,
  phi = pars$phi,
  doc_lengths = doc_lengths,
  verbose = TRUE, # verbose controls console output of progress
  threads = 1 # threads controls parallel computation
)
}
